chore(styles): move styles out of .model to .styles
