+++ packages/blockchain-wallet-v4-frontend/src/scenes/Earn/Table/Tag/Tag.styles.ts
